# baas-swagger-v7.3.yaml
swagger: "2.0"
info:
  description: "This is a basic documentation of the Blockchain as a Service (BaaS), platform that will be integrated in ioTFeds by CERTH/ITI. The baasic endpoints for user and federation magement are described here. The documentation will be adjusted and extended as the platform is integrated."
  version: "1.0.0"
  title: "On-chain information in IoTFeds"
# # host: "petstore.swagger.io"
# # basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
...
  /createProduct: # registration use case
    post:
      tags:
      - "product"
      summary: "Register product on Blockchain"
      description: "This is the process of registering a product."
      operationId: "createProduct"
      produces:
      - "message"
      parameters:
      # - in: "body"
      #   name: "body"
      #   description: "Created user object"
      #   required: true
      #   schema:
      #     $ref: "#/definitions/User" # update user definition
      - name: "product_id"
        in: "body"
        description: "The product's id"
        required: true
        type: "string"
      - name: "resource_ids"
        in: "body"
        description: "The product's related resources"
        required: true
        schema:
          type: array
          items:
            type: string
          example: ["str1", "str2", "str3"]
      - name: "product_details"
        in: "body"
        description: "The details of the product"
        required: true
        type: "object"
      - name: "seller_id"
        in: "body"
        description: "The seller/creator of the product"
        required: true
        type: "string"
      - name: "fedMarketplace_id"
        in: "body"
        description: "The id of the federated marketplace the product resides in"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful registration"
          schema:
            $ref: "#/definitions/Product" # update product definition
        "400":
          description: "registration failed"
...
  /getResourceInfo:
    get:
      tags:
      - "resource"
      summary: "Information about a resource"
      description: "Returns whether a resource exists and relative info"
      operationId: "getResourceInfo"
      produces:
      - "application/json"
      parameters:
      - name: "resource_id"
        in: "query"
        description: "The resource's id"
        required: true
        type: "string"
      responses:
        "200":
          description: "Requested resource"
          schema:
            $ref: "#/definitions/Resource" # update resource definition
        "400":
          description: "access denied"
...
  /getFederatedResources:
    get:
      tags:
      - "resource"
      summary: "Information about fedeation's resources"
      description: "Returns fedeation's resources registered in the Blockchain"
      operationId: "getFederatedResources"
      produces:
      - "application/json"
      parameters:
      - name: "user_id"
        in: "query"
        description: "The user's id"
        required: true
        type: "string"
      - name: "fed_id"
        in: "query"
        description: "The federation's id"
        required: true
        type: "string"
      - name: "Hr"
        in: "query"
        description: "Maximum upper time limit that the device is allowed to support"
        required: true
        type: "date-time"
      - name: "Lr"
        in: "query"
        description: "Minimum time limit available to support the device"
        required: true
        type: "date-time"
      - name: "Fr"
        in: "query"
        description: "The federation's id"
        required: true
        type: "number"
      - name: "price_min"
        in: "query"
        description: "The minimum price to search for"
        required: true
        type: "number"
      - name: "price_max"
        in: "query"
        description: "The maximum price to search for"
        required: true
        type: "number"
      - name: "rep_min"
        in: "query"
        description: "The minimum reputation to search for"
        required: true
        type: "number"
      - name: "rep_max"
        in: "query"
        description: "The maximum reputation to search for"
        required: true
        type: "number"
      responses:
        "200":
          description: "Federation's resources object"
          schema:
            type: array
            items:
              type: object
              $ref: "#/definitions/Resource" # update product definition
        "400":
          description: "access denied"
  /searchResources: # Serves the distributed searching for resources
      post:
        tags:
        - "resource"
        summary: "searches for resources that satisfy a set of criteria"
        description: "searches for resources that satisfy a set of criteria"
        operationId: "searchResources"
        produces:
        - "message"
        parameters:
        - name: "resources"
          in: "body"
          description: "the resources to be searched"
          required: true
          type: "array"
          items:
            type: "string"
        - name: "price_min" 
          in: "body"
          description: "The minimum price to search for"
          required: true
          type: "number"
        - name: "price_max" 
          in: "body"
          description: "The maximum price to search for"
          required: true
          type: "number"
        - name: "rep_min" 
          in: "body"
          description: "The minimum reputation to search for"
          required: true
          type: "number"
        - name: "rep_max" 
          in: "body"
          description: "The maximum reputation to search for"
          required: true
          type: "number"
        - name: "user_id" 
          in: "body"
          description: "The user's id"
          required: true
          type: "string"
        responses:
          "200":
            description: "Object with array of resources "
            schema:
              type: object
              items:
                type: array
                items:
                  resource:
                    type: "object"
              # example:
              $ref: "#/definitions/Resource"
          "403":
            description: "Search for resources failed"
          "404":
            description: "No resources found"
...
  /updateResourceQos: # serves pprice changes
    patch:
      tags:
      - "resource"
      summary: "Updates a resource's objective rep. metrics"
      description: "Updates resource's QoS"
      operationId: "updateResourceQos"
      produces:
      - "message"
      parameters:
      - name: "resource_id"
        in: "query"
        description: "The resource's id"
        required: true
        type: "string"
      - name: "qos"
        in: "query"
        description: "qos metrics, availability, response time, precision"
        required: true
        type: "object"
        properties:
          availability:
            type: "string"
          response_time:
            type: "string"
          precision:
            type: "string"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "update failed"  
...
definitions:
...
  User:
    type: "object"
    properties:
      ID:
        type: "string"
      role:
        type: "string"
      mail:
        type: "string"
      organization:
        type: "string"
      balance:
        type: "number"
      # associated_resources:
      #   type: "array"
      #   items:
      #     type: "string"
      Fed_owner:
        type: "boolean"
      Feds:
        type: array
        items:
          type: string
        example: ["federation1", "federation2"]        
      ActiveProducts:
        type: array
        items:
          type: string
        example: ["product1", "product2"]
      AssociatedPlatforms:
        type: "object"
        additionalProperties:
          type: "array"
          items:
              type: "string"
        example:
          platfrom1: ["resource1", "resource2", "resource3"]
          platfrom2: ["resource1", "resource2", "resource3"]
    xml:
      name: "User"
...
  Resource:
    type: "object"
    properties:
      Resource_id:
        type: "string"
      Platform:
        type: "string"
      Price:
        type: "number"
      subj_reputation:
        type: "number"
      overall_reputation:
        type: "number"
      platform:
        type: "string"
      docType:
        type: "string"
      owner:
        type: "string"
      federations:
        items:
          type: "string"
      obj_reputation:
        type: "number"
      transactionCounter:
        type: "number"
      Brd:
        type: "number"
      Brr:
        type: "number"
      Fr:
        type: "number"
      Hr:
        type: "number"
      Lr:
        type: "number"
      Wr:
        type: "number"
      ard:
        type: "number"
      arf:
        type: "number"
      aro:
        type: "number"
      arr:
        type: "number"
    xml:
      name: "Resource"
  Product:
    type: "object"
    properties:
      Product_id:
        type: "string"
      docType:
        type: "string"
      Seller:
        type: "string"
      FedMarketplace_id:
        type: "string"
      GlobalMarketplace:
        type: "boolean"
      Price:
        type: "number"
      subjReputation:
        type: "array"
        items:
          type: "number"
      Reputation:
        type: "number"
      Resource_ids:
        type: "object"
        items:
          type: "string"
      transactionCounter:
        type: "number"
      product_details:
        type: "array"
        items:
          type: "object"
          properties:
            product_type:
              type: "string"
            total_volume_mb:
              type: "number"
    xml:
      name: "Product"
...